+++ qiskit/extensions/standard/cu3.py
@@ -10,8 +10,8 @@ controlled-u3 gate.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.decorators import _control_target_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 from qiskit.extensions.standard.u1 import U1Gate
@@ -66,27 +66,9 @@ class Cu3Gate(Gate):
                                  self.param[2], self.qargs[0], self.qargs[1]))
 
 
+@_control_target_gate
 def cu3(self, theta, phi, lam, ctl, tgt):
     """Apply cu3 from ctl to tgt with angle theta, phi, lam."""
-    if isinstance(ctl, QuantumRegister) and \
-       isinstance(tgt, QuantumRegister) and len(ctl) == len(tgt):
-        instructions = InstructionSet()
-        for i in range(ctl.size):
-            instructions.add(self.cu3(theta, phi, lam, (ctl, i), (tgt, i)))
-        return instructions
-
-    if isinstance(ctl, QuantumRegister):
-        instructions = InstructionSet()
-        for j in range(ctl.size):
-            instructions.add(self.cu3(theta, phi, lam, (ctl, j), tgt))
-        return instructions
-
-    if isinstance(tgt, QuantumRegister):
-        instructions = InstructionSet()
-        for j in range(tgt.size):
-            instructions.add(self.cu3(theta, phi, lam, ctl, (tgt, j)))
-        return instructions
-
     self._check_qubit(ctl)
     self._check_qubit(tgt)
     self._check_dups([ctl, tgt])