+++ qiskit/extensions/standard/ry.py
@@ -12,8 +12,8 @@ Rotation around the y-axis.
 """
 from qiskit.circuit import Gate
 from qiskit.circuit import QuantumCircuit
-from qiskit.circuit import InstructionSet
 from qiskit.circuit import QuantumRegister
+from qiskit.circuit.decorators import _1q_gate
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import header  # pylint: disable=unused-import
 from qiskit.extensions.standard.u3 import U3Gate
@@ -55,14 +55,9 @@ class RYGate(Gate):
         self._modifiers(circ.ry(self.param[0], self.qargs[0]))
 
 
+@_1q_gate
 def ry(self, theta, q):
     """Apply Ry to q."""
-    if isinstance(q, QuantumRegister):
-        instructions = InstructionSet()
-        for j in range(q.size):
-            instructions.add(self.ry(theta, (q, j)))
-        return instructions
-
     self._check_qubit(q)
     return self._attach(RYGate(theta, q, self))